+++ src/components/FiledBtn.tsx
@@ -5,11 +5,12 @@ interface FiledBtnProps {
   size?: string;
   onClick?: () => void;
   cutomStyle?: SerializedStyles;
+  type?: 'button' | 'submit' | 'reset';
 }
 
-const FiledBtn = ({ children, color, size, onClick, cutomStyle }: FiledBtnProps) => {
+const FiledBtn = ({ children, color, size, onClick, cutomStyle, type }: FiledBtnProps) => {
   return (
-    <button css={[filedBtn(color, size), cutomStyle]} onClick={onClick}>
+    <button css={[filedBtn(color, size), cutomStyle]} onClick={onClick} type={type}>
       {children}
     </button>
   );
